requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
bcrypt
aiofiles
pydantic
python-multipart
python-jose
websockets